requirements.txt
fastmcp
google-genai